Revert "move files outside of the /api directory"
This reverts commit a8cdecaf150ecfcad67ba6b51c7c8824b78ff6a1.
