+++ apps/web/pages/api/contract-verification/prisma/contract-status.ts
@@ -1,6 +1,5 @@
 import { NextApiRequest, NextApiResponse } from "next";
-import prisma from "../../prisma/lib/prisma";
-
+import prisma from "./lib/prisma";
 export default async function isContractVerified(
   req: NextApiRequest,
   res: NextApiResponse
